requirements.txt
Scrapy==2.4.1
mysql-connector-python~=8.0.22
itemloaders~=1.0.4